refactor: BankClerk work(_ queue: Queue<Client>) 메서드 수정
createWorkItem 으로 workItem 분리
measureTotalWorkTime 으로 작업 시간 측정 기능 분리
Delegate 삭제
