+++ BankManagerConsoleApp/BankManagerConsoleApp/Model/BankClerk.swift
@@ -7,28 +7,26 @@
 
 import Foundation
 
-fileprivate extension Constants {
-    static let decimalPlace = "%.2f"
-}
-
 final class BankClerk: Presentable {
     private let name: String
     private let workSpeed: Double
-    private weak var delegate: BankDelegate?
     
     init(name: String, workSpeed: Double) {
         self.name = name
         self.workSpeed = workSpeed
     }
     
-    func setDelegate(delegate: BankDelegate) {
-        self.delegate = delegate
-    }
-    
     func work(_ queue: Queue<Client>) {
         let workQueue = DispatchQueue(label: name)
-        let startTime = CFAbsoluteTimeGetCurrent()
-        let workItem = DispatchWorkItem {
+        let workItem = createWorkItem(queue)
+        
+        while queue.isEmpty == false {
+            workQueue.sync(execute: workItem)
+        }
+    }
+    
+    private func createWorkItem(_ queue: Queue<Client>) -> DispatchWorkItem {
+        DispatchWorkItem {
             guard let client = queue.peek else {
                 return
             }
@@ -39,20 +37,5 @@ final class BankClerk: Presentable {
             
             queue.dequeue()
         }
-                
-        while queue.isEmpty == false {
-            workQueue.sync(execute: workItem)
-        }
-        
-        let finishTime = CFAbsoluteTimeGetCurrent()
-        let totalTime = finishTime - startTime
-        
-        delegate?.close(totalWorkTime: totalTime.formatted)
-    }
-}
-
-fileprivate extension CFAbsoluteTime {
-    var formatted: String {
-        String(format: Constants.decimalPlace, self)
     }
 }
